requirements.txt
aio-pika==9.5.7
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiormq==6.9.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.57.1
antlr4-python3-runtime==4.9.3
anyio==4.9.0
libcst
asteval==1.0.6
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.4
bidict==0.23.1
black==25.1.0
blake3==1.0.5
blinker==1.9.0
blis==1.2.1
bokeh==3.7.3
boto3==1.40.6
botocore==1.40.6
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cohere==5.6.2
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastavro==1.12.0
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.5.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.5.1
gevent==25.5.1
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.30.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.73.1
grpcio-status==1.71.2
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
hdbscan==0.8.40
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
hydra-core==1.3.2
hypothesis==6.137.1
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
llama-cloud==0.1.32
llama-cloud-services==0.6.43
llama-index==0.12.49
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.49
llama-index-embeddings-openai==0.3.1
cryptography>=42.0.0
llama-index-indices-managed-llama-cloud==0.7.10
llama-index-instrumentation==0.3.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.1.0
llama-parse==0.6.43
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
Mesa==3.2.0
mpmath==1.3.0
python-frontmatter==1.1.0
multidict==6.6.3
murmurhash==1.0.13
mypy==1.17.1
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
narwhals==2.0.1
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==2.1.0
omegaconf==2.3.0
openai==1.95.1
orjson==3.11.0
packaging==25.0
pamqp==3.3.0
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.8.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML>=6.0.1,<7
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.8
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette
striprtf==0.0.26
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.53.1
Tree==0.2.4
typer==0.16.0
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xyzservices==2025.4.0
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
zstandard==0.23.0
